# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Red Fox Hotel HRMS & Payroll API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /employees:
    get:
      operationId: getEmployees
      tags: [employees]
      summary: List employees
      parameters:
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: departmentId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Employee"
    post:
      operationId: createEmployee
      tags: [employees]
      summary: Create an employee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
  /employees/{id}:
    get:
      operationId: getEmployee
      tags: [employees]
      summary: Get an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    patch:
      operationId: updateEmployee
      tags: [employees]
      summary: Update an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
    delete:
      operationId: deleteEmployee
      tags: [employees]
      summary: Delete an employee
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # ATTENDANCE
...
  /leaves:
    get:
      operationId: getLeaveRequests
      tags: [leaves]
      summary: List leave requests
      parameters:
        - name: employeeId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaveRequest"
    post:
      operationId: createLeaveRequest
      tags: [leaves]
      summary: Apply for leave
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}:
    get:
      operationId: getLeaveRequest
      tags: [leaves]
      summary: Get a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/approve:
    patch:
      operationId: approveLeave
      tags: [leaves]
      summary: Approve a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveDecision"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"
  /leaves/{id}/reject:
    patch:
      operationId: rejectLeave
      tags: [leaves]
      summary: Reject a leave request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeaveDecision"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaveRequest"

  # ADVANCES
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /holidays:
    get:
      operationId: getHolidays
      tags: [holidays]
      summary: List holidays
      parameters:
        - name: year
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Holiday"
    post:
      operationId: createHoliday
      tags: [holidays]
      summary: Create a holiday
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HolidayInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Holiday"
  /holidays/{id}:
    delete:
      operationId: deleteHoliday
      tags: [holidays]
      summary: Delete a holiday
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No Content

  # SETTINGS
...
components:
  schemas:
...
    Employee:
      type: object
      required: [id, employeeId, firstName, lastName, email, phone, gender, department, designation, branchId, joiningDate, employmentType, status, salary]
      properties:
        id:
          type: integer
        employeeId:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        department:
          type: string
        designation:
          type: string
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        shiftId:
          type: ["integer", "null"]
        shiftName:
          type: ["string", "null"]
        weeklyOffPolicyId:
          type: ["integer", "null"]
        weeklyOffPolicyName:
          type: ["string", "null"]
        joiningDate:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract, intern]
        status:
          type: string
          enum: [active, inactive, terminated]
        salary:
          type: number
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        panNumber:
          type: ["string", "null"]
        aadhaarNumber:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    EmployeeInput:
      type: object
      required: [firstName, lastName, email, phone, gender, department, designation, branchId, joiningDate, employmentType, salary]
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          type: string
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        department:
          type: string
        designation:
          type: string
        branchId:
          type: integer
        shiftId:
          type: ["integer", "null"]
        weeklyOffPolicyId:
          type: ["integer", "null"]
        joiningDate:
          type: string
        employmentType:
          type: string
        salary:
          type: number
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        panNumber:
          type: ["string", "null"]
        aadhaarNumber:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
    EmployeeUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        gender:
          type: string
        dob:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        emergencyContact:
          type: ["string", "null"]
        department:
          type: string
        designation:
          type: string
        branchId:
          type: integer
        shiftId:
          type: ["integer", "null"]
        weeklyOffPolicyId:
          type: ["integer", "null"]
        joiningDate:
          type: string
        employmentType:
          type: string
        status:
          type: string
        salary:
          type: number
        bankName:
          type: ["string", "null"]
        accountNumber:
          type: ["string", "null"]
        ifscCode:
          type: ["string", "null"]
        upiId:
          type: ["string", "null"]
        panNumber:
          type: ["string", "null"]
        aadhaarNumber:
          type: ["string", "null"]
        photoUrl:
          type: ["string", "null"]
...
    LeaveRequest:
      type: object
      required: [id, employeeId, leaveType, startDate, endDate, status, reason]
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: ["string", "null"]
        employeeCode:
          type: ["string", "null"]
        leaveType:
          type: string
          enum: [casual, sick, earned, emergency, loss_of_pay, maternity, paternity]
        startDate:
          type: string
        endDate:
          type: string
        days:
          type: integer
        status:
          type: string
          enum: [pending, approved, rejected]
        reason:
          type: string
        managerComment:
          type: ["string", "null"]
        approvedById:
          type: ["integer", "null"]
        createdAt:
          type: string
    LeaveInput:
      type: object
      required: [employeeId, leaveType, startDate, endDate, reason]
      properties:
        employeeId:
          type: integer
        leaveType:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        reason:
          type: string
    LeaveDecision:
      type: object
      required: [comment]
      properties:
        comment:
          type: string
...
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [attendance_reminder, leave_approval, salary_released, birthday_wish, announcement]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
    Holiday:
      type: object
      required: [id, name, date]
      properties:
        id:
          type: integer
        name:
          type: string
        date:
          type: string
        createdAt:
          type: string
    HolidayInput:
      type: object
      required: [name, date]
      properties:
        name:
          type: string
        date:
          type: string
...